Update references to plugin package renaming
