+++ grpc_plugins/currency/proxy/proxy_test.go
@@ -23,8 +23,8 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/katzenpost/server_plugins/currency/common"
-	"github.com/katzenpost/server_plugins/currency/config"
+	"github.com/katzenpost/server_plugins/grpc_plugins/currency/common"
+	"github.com/katzenpost/server_plugins/grpc_plugins/currency/config"
 	"github.com/stretchr/testify/assert"
 )
